Add syntax driven symbol pickers `<space>ts`, `<space>tS`
This is roughly the same idea as the LSP symbol pickers but not as
accurate. Prior art is the GitHub search-based navigation:
<https://docs.github.com/en/repositories/working-with-files/using-files/navigating-code-on-github#about-navigating-code-on-github>

They use `tags.scm` queries but I think we will want to extend these
so I've added a new `symbol.scm` query file.

Why do this at all?

* Some LSs are slow
    * I even wrote a tiny language server that uses tree-sitter the
      same way as this because ErlangLS has a slow boot time and I
      usually hit `<space>s` as my first action in a file.
* Some LSs don't implement features like workspace search in a useful
  way.
    * For example Erlang language servers like ErlangLS and ELP only
      show module names.
* Some that are good can be hard to configure or use on existing
  projects, like clangd on big/old C codebases.
    * For example the C sources of Erlang/OTP. `clangd` is a hinderance
      because of diagnostics in that repo.
* Sometimes you don't want install a languge server just to read
  through a codebase. When reading, all I want is symbol search
  (document & workspace) and goto-definition/reference.

So this commit adds a symbol picker for the current document that
extracts symbol information using a `symbol.scm` query, and also a
picker for the workspace.

The workspace picker works by scanning the symbols for any open
documents and then running a search using WalkBuilder and grep-searcher,
just like `global_search`. Any paths that aren't open yet but match
the regex are opened, parsed and scanned for matching symbols.